Fix `ark --install` on Linux
And fix `--connection_file` error
