+++ crates/ark/src/version.rs
@@ -10,7 +10,7 @@ use std::env;
 use std::path::PathBuf;
 
 use anyhow::Context;
-use harp::command::r_command;
+use harp::command::r_command_from_path;
 use harp::object::RObject;
 use itertools::Itertools;
 use libr::SEXP;
@@ -30,19 +30,35 @@ pub struct RVersion {
     pub r_home: String,
 }
 
+pub(crate) fn r_home_setup() -> PathBuf {
+    match std::env::var("R_HOME") {
+        Ok(home) => {
+            // Get `R_HOME` from env var, typically set by Positron / CI / kernel specification
+            PathBuf::from(home)
+        },
+        Err(_) => {
+            // Get `R_HOME` from `PATH`, via `R`
+            let Ok(result) = r_command_from_path(|command| {
+                command.arg("RHOME");
+            }) else {
+                panic!("Can't find R or `R_HOME`");
+            };
+
+            let r_home = String::from_utf8(result.stdout).unwrap();
+            let r_home = r_home.trim();
+
+            // Now set `R_HOME`. From now on, `r_command()` can be used to
+            // run exactly the same R as is running in Ark.
+            unsafe { std::env::set_var("R_HOME", r_home) };
+            PathBuf::from(r_home)
+        },
+    }
+}
+
 pub fn detect_r() -> anyhow::Result<RVersion> {
-    let output = r_command(|command| {
-        command.arg("RHOME");
-    })
-    .context("Failed to execute R to determine R_HOME")?;
+    let r_home: String = r_home_setup().to_string_lossy().to_string();
 
-    // Convert the output to a string
-    let r_home = String::from_utf8(output.stdout)
-        .context("Failed to convert R_HOME output to string")?
-        .trim()
-        .to_string();
-
-    let output = r_command(|command| {
+    let output = r_command_from_path(|command| {
         command
             .arg("--vanilla")
             .arg("-s")
